+++ src/sdk/PnP.Core/Model/Teams/Internal/TeamChatMessage.cs
@@ -1,8 +1,5 @@
 ﻿using PnP.Core.Services;
 using System;
-using System.Collections.Generic;
-using System.Dynamic;
-using System.Text.Json;
 using System.Threading.Tasks;
 
 namespace PnP.Core.Model.Teams