+++ src/sdk/PnP.Core/Model/Teams/Public/ITeamChatMessageReplyCollection.cs
@@ -1,7 +1,5 @@
-﻿using PnP.Core.Services;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 
 namespace PnP.Core.Model.Teams
 {